+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @perimetre/tokens
 
+## 0.3.0
+
+### Minor Changes
+
+- 496ae71: Changing p tag into span tag in eyebrow field
+- 496ae71: Created SectionHorizontalHeader component
+
+## 0.2.0
+
+### Minor Changes
+
+- 00b3faa: Created SectionHorizontalHeader component
+
 ## 0.1.0
 
 ### Minor Changes